requirements.txt
uagents
serpapi
streamlit
pandasai
langchain_community
langchain
datasets
python-dotenv
seaborn
ipywidgets